#20 404페이지 UI 수정
